mise en place des protections des routes, des middlewares, et la vue.

+++ blog/app/Http/Controllers/ArticlesController.php
@@ -2,12 +2,26 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Article;
 use App\Models\User;
+use App\Models\Article;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controllers\Middleware;
+use Illuminate\Routing\Controllers\HasMiddleware;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
+use Illuminate\Support\Facades\Auth;
 
-class ArticlesController extends Controller
+class ArticlesController extends Controller implements HasMiddleware
 {
+
+    use AuthorizesRequests; // Incluez le trait pour avoir accès à authorize()
+
+    public static function middleware(): array
+    {
+        // Déclarer ici les middlewares pour ce contrôleur
+        return [
+            new Middleware('auth', except: ['index', 'show']),
+        ];
+    }
     public function index()
     {
         $articles = Article::with('user')->orderBy('created_at', 'desc')->get();
@@ -31,13 +45,15 @@ class ArticlesController extends Controller
 
     public function create()
     {
+        $this->authorize('create', Article::class);
         return view('articles.create'); // Retourne la vue de création d'article
     }
 
     public function store(Request $request)
     {
+        $this->authorize('create', Article::class);
         // vérification des permissions plus tard
-        $user = User::find(1);
+        $user = Auth::user(); //On récupère l'utilisateur connecté
         $request['user_id'] = $user->id;
 
         $validatedData = $request->validate([
@@ -57,17 +73,20 @@ class ArticlesController extends Controller
 
     public function edit(Article $article)
     {
+        $this->authorize('update', $article);
         return view('articles.edit', compact('article'));
     }
 
     public function update(Request $request, Article $article)
     {
+        $this->authorize('update', $article);
         $article->update($request->all());
         return redirect('/articles');
     }
 
     public function delete(Article $article)
     {
+        $this->authorize('delete', $article);
         // vérification des permissions plus tard
         $article->delete();
         return redirect('/articles');
